restructure code; add accessor functions to pipeline.
